requirements.txt
fastapi
uvicorn
httpx
termcolor
jinja2
python-dotenv
openai
sqlalchemy
psycopg2-binary